+++ packages/commonwealth/client/scripts/views/pages/Editor/indicators/DragIndicator.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/shared';
+@import '../../../../../styles/shared';
 
 .DragIndicator {
   position: absolute;
